fix: Update the type annotation of `retry_seconds` in the state backend `acquire_lock` method to allow `float` values (#9921)
## Description

<!-- Describe the changes introduced by this PR -->

## Checklist

- [x] I have read the [contribution
guide](https://docs.meltano.com/contribute/merge/)

### Was generative AI tooling used to co-author this PR?

<!--
If generative AI tooling has been used in the process of authoring this
PR, please change below checkbox to `[X]` followed by the name of the
tool, uncomment the "Generated-by". See the agentic coding section of
the contribution guide for details:
https://github.com/meltano/meltano/blob/main/CONTRIBUTING.md#agentic-coding
-->

- [ ] Yes (please specify the tool below)

<!--
Generated-by: [Tool Name] following [the agentic coding
guidelines](https://github.com/meltano/meltano/blob/main/CONTRIBUTING.md#agentic-coding)
-->

## Related Issues

* Closes #XXXX

## Summary by Sourcery

Allow fractional retry intervals in state backend lock acquisition and
align type hints and overrides across implementations.

Bug Fixes:
- Permit float values for the retry_seconds parameter in state store
acquire_lock methods to match actual usage.

Enhancements:
- Annotate state store manager methods with override for better type
checking and maintenance.

Documentation:
- Update custom state backend guide to document retry_seconds as a float
parameter in acquire_lock.

+++ src/meltano/core/state_store/filesystem.py
@@ -268,7 +268,7 @@ class BaseFilesystemStateStoreManager(StateStoreManager):
         self,
         state_id: str,
         *,
-        retry_seconds: int,
+        retry_seconds: float,
     ) -> Generator[None, None, None]:
         """Context manager for locking state_id during reads and writes.
